+++ packages/browser/lib/main.ts
@@ -5,7 +5,7 @@ import {
   buildCageKeyFromSuppliedPublicKey,
   deriveSharedSecret,
 } from "./utils";
-import Config, { ConfigUrls } from "./config";
+import Config, { ConfigUrls, SdkContext } from "./config";
 import { CoreCrypto, Http, Forms, Input } from "./core";
 import { base64StringToUint8Array } from "./encoding";
 
@@ -51,11 +51,16 @@ export default class EvervaultClient {
       customConfig?.publicKey
     );
 
+    const context = this.getContext(
+      window?.location?.origin ?? "",
+      this.config.input.inputsUrl
+    );
+
     this.http = Http(
       this.config.http,
       this.config.teamId,
       this.config.appId,
-      this.config.input.inputsUrl ? "inputs" : "default"
+      context
     );
 
     this.forms = Forms(this);
@@ -102,6 +107,14 @@ export default class EvervaultClient {
       this.config.encryption,
       this.isInDebugMode()
     );
+  }
+
+  getContext(origin: string, inputsUrl: string): SdkContext {
+    if (origin === inputsUrl) {
+      return "inputs";
+    } else {
+      return "default";
+    }
   }
 
   /**
@@ -154,6 +167,28 @@ export default class EvervaultClient {
     } catch (err) {
       throw err;
     }
+  }
+
+  reveal(
+    elementId: string,
+    request: Request,
+    config?: Record<string, any>
+  ): {
+    isRevealLoaded: Promise<boolean>;
+  } {
+    let reveal = this.input.generate(
+      elementId,
+      {
+        theme: "reveal",
+        ...config,
+      },
+      true,
+      request
+    );
+
+    return {
+      isRevealLoaded: reveal.isInputsLoaded,
+    };
   }
 
   /**