+++ e2e-apps/decrypt-backend/CHANGELOG.md
@@ -35,3 +35,27 @@
 ### Patch Changes
 
 - @evervault/browser-e2e-app@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null