+++ e2e-tests/browser/CHANGELOG.md
@@ -41,3 +41,31 @@
 
 - @evervault/browser-e2e-app@null
 - @evervault/e2e-decrypt-backend@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null
+- @evervault/e2e-decrypt-backend@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null
+- @evervault/e2e-decrypt-backend@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null
+- @evervault/e2e-decrypt-backend@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null
+- @evervault/e2e-decrypt-backend@null